// src/types/lib.d.ts
export interface ZKIRootState {
  accountModalOpened: boolean;
  currentModal?: string;
  previousRoute?: Route;
  user?: unknown,
  authenticated: boolean,
}